make separate class structure file
